requirements.txt
# Requirements for the FastAPI application with OCR capabilities
fastapi
streamlit
uvicorn
pillow
easyocr
requests
python-multipart
